Add cronjob for auto publishing cache flush
Usage:
typo3/cli_dispatch.phpsh extbase autoflush:clearmenuforpulishedpages
or launch from EXT:scheduler

The cronjob stores the last run timestamp in the cache_hash cache
and clears pages for all pages that published/or ended between last
and current run.

This clears the menu_pid_UID tags, which need to be set in frontend
rendering using the RegisterMenu Helper.
See Hooks/MenuFlush.php for more info.

Note: It would be better to fix this in core and have a method to define cache expiry
runtime.
But the HMENU relies on a static cache expiration calculation.
That means the cache expiration can't be calculated from the rendered content,
Fixing this in core means to rewrite the cache handling of HMENU.
Therefore we implement this as cronjob for now.